refactor: organize use cases into feature-specific packages and update imports
