+++ cli/src/main/kotlin/com/github/adriianh/cli/command/player/SearchCommand.kt
@@ -5,15 +5,15 @@ import com.github.adriianh.cli.config.configDir
 import com.github.adriianh.cli.config.resolveEnv
 import com.github.adriianh.cli.di.appModule
 import com.github.adriianh.core.domain.model.Track
-import com.github.adriianh.core.domain.usecase.GetLyricsUseCase
-import com.github.adriianh.core.domain.usecase.GetSimilarTracksUseCase
-import com.github.adriianh.core.domain.usecase.SearchTracksUseCase
+import com.github.adriianh.core.domain.usecase.search.GetLyricsUseCase
+import com.github.adriianh.core.domain.usecase.search.GetSimilarTracksUseCase
+import com.github.adriianh.core.domain.usecase.search.SearchTracksUseCase
 import com.github.ajalt.clikt.core.CliktCommand
 import com.github.ajalt.clikt.core.Context
 import com.github.ajalt.clikt.parameters.arguments.argument
+import com.github.ajalt.clikt.parameters.options.default
 import com.github.ajalt.clikt.parameters.options.flag
 import com.github.ajalt.clikt.parameters.options.option
-import com.github.ajalt.clikt.parameters.options.default
 import com.github.ajalt.clikt.parameters.types.choice
 import com.github.ajalt.clikt.parameters.types.int
 import kotlinx.coroutines.runBlocking
