+++ cli/src/main/kotlin/com/github/adriianh/cli/di/AppModule.kt
@@ -11,7 +11,18 @@ import com.github.adriianh.core.domain.provider.AudioProvider
 import com.github.adriianh.core.domain.provider.DiscoveryProvider
 import com.github.adriianh.core.domain.provider.MusicProvider
 import com.github.adriianh.core.domain.repository.*
-import com.github.adriianh.core.domain.usecase.*
+import com.github.adriianh.core.domain.usecase.library.*
+import com.github.adriianh.core.domain.usecase.offline.*
+import com.github.adriianh.core.domain.usecase.playback.*
+import com.github.adriianh.core.domain.usecase.search.*
+import com.github.adriianh.core.domain.usecase.session.ClearSessionUseCase
+import com.github.adriianh.core.domain.usecase.session.RestoreSessionUseCase
+import com.github.adriianh.core.domain.usecase.session.SaveSessionUseCase
+import com.github.adriianh.core.domain.usecase.settings.GetSettingsUseCase
+import com.github.adriianh.core.domain.usecase.settings.UpdateSettingsUseCase
+import com.github.adriianh.core.domain.usecase.stats.GetListeningStatsUseCase
+import com.github.adriianh.core.domain.usecase.stats.GetTopArtistsUseCase
+import com.github.adriianh.core.domain.usecase.stats.GetTopTracksUseCase
 import com.github.adriianh.data.local.DatabaseFactory
 import com.github.adriianh.data.local.MeloDatabase
 import com.github.adriianh.data.provider.artwork.CompositeArtworkProvider
