+++ core/src/main/kotlin/com/github/adriianh/core/domain/usecase/search/GetSyncedLyricsUseCase.kt
@@ -1,12 +1,12 @@
-package com.github.adriianh.core.domain.usecase
+package com.github.adriianh.core.domain.usecase.search
 
 import com.github.adriianh.core.domain.repository.LyricsRepository
 
-class GetLyricsUseCase(
+class GetSyncedLyricsUseCase(
     private val repository: LyricsRepository
 ) {
     suspend operator fun invoke(artist: String, title: String): String? {
         if (artist.isBlank() || title.isBlank()) return null
-        return repository.getLyrics(artist, title)
+        return repository.getSyncedLyrics(artist, title)
     }
 }
